+++ pkg/clients/mocks/helm.go
@@ -8,14 +8,20 @@ type MockHelm struct {
 	mock.Mock
 }
 
-func (h *MockHelm) Create(kubeConfig, name, namespace string, createNamespace bool, chartPath, valuesPath string, valueString map[string]string) error {
-	args := h.Called(kubeConfig, name, namespace, createNamespace, chartPath, valuesPath, valueString)
+func (h *MockHelm) Create(kubeConfig, name, namespace string, createNamespace bool, chart, version, valuesPath string, valueString map[string]string) error {
+	args := h.Called(kubeConfig, name, namespace, createNamespace, chart, version, valuesPath, valueString)
 
 	return args.Error(0)
 }
 
 func (h *MockHelm) Destroy(kubeConfig, name, namespace string) error {
 	args := h.Called(kubeConfig, name, namespace)
+
+	return args.Error(0)
+}
+
+func (h *MockHelm) UpsertChartRepository(name, url string) error {
+	args := h.Called(name, url)
 
 	return args.Error(0)
 }