+++ pkg/clients/system.go
@@ -2,7 +2,6 @@ package clients
 
 import (
 	"bufio"
-	"context"
 	"fmt"
 	"io"
 	"io/ioutil"
@@ -11,8 +10,6 @@ import (
 	"runtime"
 	"strings"
 	"time"
-
-	"github.com/docker/docker/api/types"
 )
 
 const (
@@ -73,6 +70,7 @@ func (b *SystemImpl) OpenBrowser(uri string) error {
 // working correctly
 func (b *SystemImpl) Preflight() (string, error) {
 	dockerPass := true
+	podmanPass := true
 	gitPass := true
 	errors := ""
 	output := ""
@@ -85,6 +83,14 @@ func (b *SystemImpl) Preflight() (string, error) {
 		dockerPass = false
 	} else {
 		output += fmt.Sprintf(" [ %s ] Docker\n", fmt.Sprintf(Green, "  OK   "))
+	}
+
+	if checkPodman() != nil {
+		output += fmt.Sprintf(" [ %s ] Podman\n", fmt.Sprintf(Red, " ERROR "))
+		errors += "* Unable to connect to Podman, ensure Podman is installed and running.\n"
+		podmanPass = false
+	} else {
+		output += fmt.Sprintf(" [ %s ] Podman\n", fmt.Sprintf(Green, "  OK   "))
 	}
 
 	if checkGit() != nil {
@@ -104,7 +110,7 @@ func (b *SystemImpl) Preflight() (string, error) {
 		}
 	}
 
-	if !dockerPass || !gitPass {
+	if (!dockerPass && !podmanPass) || !gitPass {
 		return fmt.Sprintf("%s\n\n%s", output, errors), fmt.Errorf("Errors preflighting system")
 	}
 
@@ -173,9 +179,33 @@ func checkDocker() error {
 		return err
 	}
 
-	_, err = d.ContainerList(context.Background(), types.ContainerListOptions{All: true})
+	dt := NewDockerTasks(d, nil, nil, nil)
+
+	if dt == nil {
+		return fmt.Errorf("unable to determine docker engine, please check that Docker or Podman is installed and the DOCKER_HOST is set")
+	}
+
+	// check that the server is a docker engine not podman
+	// if Docker there will be a component called "Engine"
+	if dt.EngineType != EngineTypeDocker {
+		return fmt.Errorf("Platform is not Docker")
+	}
+
+	return nil
+}
+
+func checkPodman() error {
+	d, err := NewDocker()
 	if err != nil {
 		return err
+	}
+
+	dt := NewDockerTasks(d, nil, nil, nil)
+
+	// check that the server is a docker engine not podman
+	// if Docker there will be a component called "Engine"
+	if dt.EngineType != EngineTypePodman {
+		return fmt.Errorf("Platform is not Podman")
 	}
 
 	return nil