+++ pkg/providers/cluster_k3s.go
@@ -24,7 +24,7 @@ import (
 // https://github.com/rancher/k3d/blob/master/cli/commands.go
 
 const k3sBaseImage = "shipyardrun/k3s"
-const k3sBaseVersion = "v1.22.3"
+const k3sBaseVersion = "v1.22.4"
 
 var startTimeout = (300 * time.Second)
 
@@ -242,7 +242,8 @@ func (c *K8sCluster) createK3s() error {
 		return err
 	}
 
-	err = c.kubeClient.HealthCheckPods([]string{""}, startTimeout)
+	// ensure essential pods have started before announcing the resource is available
+	err = c.kubeClient.HealthCheckPods([]string{"app=local-path-provisioner", "k8s-app=kube-dns"}, startTimeout)
 	if err != nil {
 		// fetch the logs from the container before exit
 		lr, lerr := c.client.ContainerLogs(id, true, true)
@@ -284,7 +285,7 @@ func (c *K8sCluster) waitForStart(id string) error {
 		out, err := c.client.ContainerLogs(id, true, true)
 		if err != nil {
 			out.Close()
-			return fmt.Errorf(" Couldn't get docker logs for %s\n%+v", id, err)
+			return fmt.Errorf("Couldn't get docker logs for %s\n%+v", id, err)
 		}
 
 		// read from the log and check for Kublet running
@@ -464,6 +465,11 @@ func (c *K8sCluster) ImportLocalDockerImages(name string, id string, images []co
 	imgs := []string{}
 
 	for _, i := range images {
+		// do nothing when the image name is empty
+		if i.Name == "" {
+			continue
+		}
+
 		err := c.client.PullImage(i, false)
 		if err != nil {
 			return err
@@ -500,14 +506,6 @@ func (c *K8sCluster) destroyK3s() error {
 	}
 
 	for _, i := range ids {
-		// remove from the networks
-		for _, n := range c.config.Networks {
-			err := c.client.DetachNetwork(n.Name, i)
-			if err != nil {
-				return err
-			}
-		}
-
 		err := c.client.RemoveContainer(i, false)
 		if err != nil {
 			return err